requirements.txt
paho-mqtt==1.6.1
pytest==8.4.1
pytest-mock==3.14.1
coverage==7.9.1
pytest-cov==6.2.1